pyproject.toml
[project]
name = "prefix-rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.59.0",
    "cairosvg>=2.8.2",
    "datasets>=3.6.0,<4.0.0",
    "dreamsim>=0.2.1",
    "google-genai>=1.27.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.9.4",
    "modal>=1.1.0",
    "numpy>=2.0.2",
    "omegaconf>=2.3.0",
    "openai>=1.82.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.55.0",
    "trl>=0.19.1",
    "wandb>=0.21.0",
]

[dependency-groups]
dev = ["pytest>=8.4.1"]

[tool.pytest.ini_options]
testpaths = ["tests"]
